allow playbooks to specify stdin for command
This commit exposes the `data` and `binary_data` arguments of
`module.run_command` to the `command` module, so that playbooks can
provide commands with input on stdin.  For example, this can be used to
embed certain types of configuration information directly into the
playbook:

    - command: virsh net-define /dev/stdin
      args:
        data: |
          <network>
            [...]
          </network>

This is often simpler than copying over a temporary file just to feed
it to a command.